Revision summary:
-Cleanups / fixes
-removed warnings
+Fully embedded the new logging system
+Added NO_VERIFY def
+Added optimized build configuration

How to compile the optimized build:
Open with your favorite text editor the following files :

/nullDC/types.h
/nullDC/log/logging_interface.h

Look for : "#undef NO_VERIFY" (in types.h) and replace it with : #define NO_VERIFY
Look for : "#define DEFAULT_LOGGING" (logging_interface.h) and replace it with : #undef DEFAULT_LOGGING

Set build settings to "Release-NoTrace" then , save everything and compile.

If you do the above , you'll notice a big difference in the block compile times and the overall performance of
the emulator.

:| |:

git-svn-id: http://nulldc.googlecode.com/svn/trunk@109 982dbc68-9c94-0dc9-9c4e-6a7e43fe6087