Revision changelog: -Improved the "Shenmue tree fix" and now the trees get a nice blending. -Improved plant rendering in Shenmue +Added some comments in the renderer to explain how the above fix works +Added fast-forward key (page up) +Added define "INCLUDE_DEV_TOOLS" -Updated compiling-advanced.txt
Revision notes:
The "FFWD key" is not meant to be a cheat or w/e
since it will mess up audio sync to the point the game re-initializes
everything(ie during the load of a level/section).

It is there to serve as dev tool.

Reviion attachments:

Smooth & blended trees :
http://img716.imageshack.us/img716/2808/shentreesalphablended1.png
http://img404.imageshack.us/img404/7641/shentreesalphablended2.png

Plant rendering improved:
http://img121.imageshack.us/img121/148/shenplantsimp.png

git-svn-id: http://nulldc.googlecode.com/svn/trunk@113 982dbc68-9c94-0dc9-9c4e-6a7e43fe6087